Revert "refactor(connlib): split device handler for clients & gateway (#2301)"
This reverts commit dbf0e445b00ebf375bd5c68dffc5e1752f2e5dff.
