+++ rust/connlib/tunnel/src/ip_packet.rs
@@ -32,20 +32,10 @@ macro_rules! swap_src_dst {
 impl<'a> MutableIpPacket<'a> {
     #[inline]
     pub(crate) fn new(data: &mut [u8]) -> Option<MutableIpPacket> {
-        let packet = match data[0] >> 4 {
-            4 => MutableIpv4Packet::new(data)?.into(),
-            6 => MutableIpv6Packet::new(data)?.into(),
-            _ => return None,
-        };
-
-        Some(packet)
-    }
-
-    #[inline]
-    pub(crate) fn destination(&self) -> IpAddr {
-        match self {
-            MutableIpPacket::MutableIpv4Packet(i) => i.get_destination().into(),
-            MutableIpPacket::MutableIpv6Packet(i) => i.get_destination().into(),
+        match data[0] >> 4 {
+            4 => MutableIpv4Packet::new(data).map(Into::into),
+            6 => MutableIpv6Packet::new(data).map(Into::into),
+            _ => None,
         }
     }
 
@@ -94,13 +84,6 @@ impl<'a> MutableIpPacket<'a> {
         match self {
             Self::MutableIpv4Packet(p) => p.to_immutable().into(),
             Self::MutableIpv6Packet(p) => p.to_immutable().into(),
-        }
-    }
-
-    pub(crate) fn as_immutable(&self) -> IpPacket<'_> {
-        match self {
-            Self::MutableIpv4Packet(p) => IpPacket::Ipv4Packet(p.to_immutable()),
-            Self::MutableIpv6Packet(p) => IpPacket::Ipv6Packet(p.to_immutable()),
         }
     }
 
@@ -191,6 +174,14 @@ pub(crate) enum IpPacket<'a> {
 }
 
 impl<'a> IpPacket<'a> {
+    pub(crate) fn new(data: &[u8]) -> Option<IpPacket> {
+        match data[0] >> 4 {
+            4 => Ipv4Packet::new(data).map(Into::into),
+            6 => Ipv6Packet::new(data).map(Into::into),
+            _ => None,
+        }
+    }
+
     pub(crate) fn version(&self) -> Version {
         match self {
             IpPacket::Ipv4Packet(_) => Version::Ipv4,
@@ -223,17 +214,17 @@ impl<'a> IpPacket<'a> {
             .flatten()
     }
 
-    pub(crate) fn source(&self) -> IpAddr {
-        match self {
-            Self::Ipv4Packet(p) => p.get_source().into(),
-            Self::Ipv6Packet(p) => p.get_source().into(),
-        }
-    }
-
     pub(crate) fn destination(&self) -> IpAddr {
         match self {
             Self::Ipv4Packet(p) => p.get_destination().into(),
             Self::Ipv6Packet(p) => p.get_destination().into(),
+        }
+    }
+
+    pub(crate) fn source(&self) -> IpAddr {
+        match self {
+            Self::Ipv4Packet(p) => p.get_source().into(),
+            Self::Ipv6Packet(p) => p.get_source().into(),
         }
     }
